Move MRT-only tests to `test_mrt.py` instead of `test_cds.py`
